experiments/mcp-gke/pyproject.toml
[project]
name = "mcp-gke"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google>=3.0.0",
    "google-genai>=1.21.1",
    "kubernetes>=33.1.0",
    "mcp>=1.9.4",
    "pyyaml>=6.0.2",
]
